purging lost files is modified that it first tries to find the lost files by looking if a file with same filename (without path and extension) and same play length exists in the database, or a song with same title artist and album and play time exists. a file scan and db update ios always done before purging. purging is not selectable alone anymore.
fix bug that full list is refreshed on song change , noticable on gtk and osx.
fix bug that if playing ends, the last song was still displayed in bold font.

+++ trunk/wxMusik/src/Frames/MusikAutoTaggerFrame.cpp
@@ -221,7 +221,7 @@ void CMusikAutoTaggerFrame::OnBnRemoveMask( wxCommandEvent& event )
 void CMusikAutoTaggerFrame::OnOk( wxCommandEvent& event )
 {
 	wxGetApp().Prefs.sAutoTag.Empty();
-	for(size_t i = 0 ; i < m_CBTagMask->GetCount();i++)
+	for(int i = 0 ; i < m_CBTagMask->GetCount();i++)
 	{
 		wxGetApp().Prefs.sAutoTag += m_CBTagMask->GetString(i);
 		wxGetApp().Prefs.sAutoTag +='|';
